+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v5.0.0] - 2025-07-22
+
+### Added
+
+- AWS trace ID now added to laravel Context facade for tracing/logging
+- Add support for Laravel 12
+
+### Removed
+
+- Custom LogstashFormatter
+- Dropped support for Laravel 10
+- Dropped support for PHP 8.1
+
 ## [v4.1.0] - 2024-11-29
 
 ### Added